03-higher-ed-text-analytics/pyproject.toml
[project]
name = "higher-ed-text-analytics"
version = "0.1.0"
description = "Higher Education Policy Text Analytics — BERTopic modeling, sentiment analysis, and keyword extraction on education research documents."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12,<3.14"
authors = [{ name = "Ruthvik Bandari" }]
keywords = ["nlp", "higher-education", "bertopic", "sentiment", "text-analytics"]

dependencies = [
    "polars[all]>=1.38",
    "pyarrow>=18.0",
    "bertopic>=0.16",
    "sentence-transformers>=3.0",
    "keybert>=0.8",
    "spacy>=3.8",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "plotly>=6.0",
    "kaleido>=0.4",
    "seaborn>=0.13",
    "matplotlib>=3.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "spacy-lookups-data>=1.0.5",
    "en-core-web-sm",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "ipykernel>=6.29",
    "nbformat>=5.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF", "N", "D"]
ignore = ["D100", "D104", "D102", "RUF001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
